+++ internal/fnos/process_windows.go
@@ -1,0 +1,11 @@
+//go:build windows
+
+package fnos
+
+import "os"
+
+func stopProcess(process *os.Process) {
+	if process != nil {
+		_ = process.Kill()
+	}
+}